+++ CHANGELOG.md
@@ -4,15 +4,24 @@ All Notable changes to `League\Period` will be documented in this file
 ## 2.5.0 - XXXX-XX-XX
 
 ### Added
-- `TimeRangeInterface`
-- `PeriodInterface`
-- `TimeRanageInterface::getDatePeriod` method
+
+- Now Period implements the following Interfaces
+    - `TimeRange`
+    - `TimeRangeInfo`
+    - `TimeRangeObject`
+
+- `Period::getDatePeriod`
+- `Period::createFromTimeRange`
 
 ### Fixed
-- DateInterval when created from an integer.
+- `DateInterval` validation when created from an integer.
 
 ### Deprecated
 - `Period::getRange` you should use `TimeRanageInterface::getDatePeriod` instead
+
+### Remove
+- support for PHP 5.3
+
 
 ## 2.4.0 - 2014-12-18